[Home/Artist] Implement correct follow/replace animation.
In the correct design, the two cards are never on screen at the same
time, this made the animation significantly simpler. After the followed
artist card is removed from the screen, the rail is re-rendered with the
suggested card taking the place of the followed artist card, which is
then animated in place.