Fixes a problem where the app's back button would become misaligned.
This would happen when the City Guides controller was in-memory (even if it wasn't in the view hierarchy). It was still getting the `viewWillAppear:` callback and setting state on the ARTopMenuViewController singleton. The fix was to only invoke the callback to update the state if the view has a window (ie: if it is in the on-screen view hierarchy).