Restricts git diff range.
This range is used to skip native CI on pull requests that have no native code changes. What was happening was, if native code got changed on master, that would "count" as a changed native file on every PR (unless they rebased). This change restricts the git diff to the most common ancestor commit of the PR's branch (ie: the commit it branched off of master from).