fix: Provide complete filter options for Works by Artists I Follow grid
When navigating from the new fair screen to the "Works by Artists I Follow"
artwork grid, we set an initial filter. This initial filter causes the filter
options (backed by aggregations) to also be scoped to the initial filter. A
user can disable the initial filter, but the aggregations will not update to
reflect the complete set of available filters for the unscoped artworks.

This commit prefetches the artwork aggregations and passes them into the
artwork grid component.